# apps/releases/var/www/api/app/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "500762 HDMI Digital Signage Media Player Api Documentation",
        "version": "1.0.0"
    },
    "host": "device_ip:8888",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/signin": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User login",
                "description": "basic authentication by token, 1 hour idle token expiration.",
                "operationId": "signin",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "token": {
                                                "type": "string",
                                                "example": "963c6f33e8d7643eb5301c6c630effce"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "authentication failed",
                        "schema": {
                            "$ref": "#/definitions/401_response"
                        }
                    }
                }
            }
        },
        "/auth/change-password": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Change user password",
                "description": "Change user password",
                "operationId": "change-password",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token retrieved from the login API",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "required": [
                                "password_old",
                                "password"
                            ],
                            "properties": {
                                "password_old": {
                                    "type": "string",
                                    "example": "admin2"
                                },
                                "password": {
                                    "type": "string",
                                    "example": "admin"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "token": {
                                                "type": "string",
                                                "example": "963c6f33e8d7643eb5301c6c630effce"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "validation failed - OBS: the code response is 200 but swagger does not allow multiple response schemas for the same code",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "failed"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "password_old": {
                                                "type": "string",
                                                "example": "Old password does not match"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "authentication failed",
                        "schema": {
                            "$ref": "#/definitions/401_response"
                        }
                    }
                }
            }
        },
        "/auth/signout": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User logout",
                "description": "User logout.",
                "operationId": "sighout",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Access token retrieved from the login API",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "authentication failed",
                        "schema": {
                            "$ref": "#/definitions/401_response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "401_response": {
            "properties": {
                "status": {
                    "type": "string",
                    "example": "failed"
                },
                "data": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "UserLogin": {
            "required": [
                "username",
                "password"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "example": "admin"
                },
                "password": {
                    "type": "string",
                    "example": "admin"
                }
            },
            "type": "object",
            "xml": {
                "name": "UserLogin"
            }
        }
    },
...
}